2024-08-21-AddStatusCodes

+++ PelicanManagement.Application/Services/Implementations/ManagementService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Identity;
+using Newtonsoft.Json;
 using PelicanManagement.Application.Security;
 using PelicanManagement.Application.Services.Interfaces;
 using PelicanManagement.Domain.Dtos.Common;
@@ -26,7 +27,7 @@ namespace PelicanManagement.Application.Services.Implementations
 {
     public class ManagementService : IManagementService
     {
-        private readonly IIdentityServerRepository _identityServer;
+        private readonly IIdentityServerRepository _identityServerRepository;
         private readonly ILogService _logService;
 
         private readonly IPelicanRepository _pelicanRepository;
@@ -39,9 +40,9 @@ namespace PelicanManagement.Application.Services.Implementations
             IPasswordHasher<Domain.Entities.IdentityServer.User> passwordHasher,
             IPelicanGenericRepository<Domain.Entities.Pelican.UsersUnit> unitGenericRepository,
             IPelicanGenericRepository<Domain.Entities.Pelican.UserPermission> permissionGenericRepository,
-            IMapper mapper, IPelicanRepository pelicanRepository,ILogService logService)
+            IMapper mapper, IPelicanRepository pelicanRepository, ILogService logService)
         {
-            _identityServer = identityServerRepository;
+            _identityServerRepository = identityServerRepository;
             _pelicanRepository = pelicanRepository;
             _mapper = mapper;
             _passwordHasher = passwordHasher;
@@ -50,17 +51,17 @@ namespace PelicanManagement.Application.Services.Implementations
             _logService = logService;
         }
 
-        public async Task<ResponseDto<bool>> AddUser(AddIdentityUserDto request, Guid operatorId)
+        public async Task<ResponseDto<bool>> AddUser(AddIdentityUserDto request, Guid operatorId, UserType type)
         {
-            var isExist = await _identityServer.IsExist(x => x.UserName.Equals(request.UserName) || x.PersonalCode.Equals(request.PersonalCode));
+            var isExist = await _identityServerRepository.IsExist(x => x.UserName.Equals(request.UserName) || x.PersonalCode.Equals(request.PersonalCode));
             if (isExist)
             {
                 return new ResponseDto<bool> { IsSuccessFull = false, Message = ErrorsMessages.RecordAlreadyExists };
             }
-
+            /// Add To Identity 
             var mappedUser = _mapper.Map<Domain.Entities.IdentityServer.User>(request);
             mappedUser.PasswordHash = _passwordHasher.HashPassword(mappedUser, request.Password);
-            await _identityServer.AddAsync(mappedUser);
+            await _identityServerRepository.AddAsync(mappedUser);
 
             if (request.PermissionIds != null)
             {
@@ -78,45 +79,78 @@ namespace PelicanManagement.Application.Services.Implementations
                     await _unitGenericRepository.AddAsync(userUnits);
                 }
             }
-            await _logService.InsertUserActivityLog(new UserActivityLogDto { UserId = operatorId, NewValues = mappedUser.UserName, UserActivityLogTypeId = ActivityLogType.CreatePelicanUser });
+
+            /// Add To Pelican 
+
+            if (type == UserType.Pelican)
+            {
+                var mappedPelicanUser = _mapper.Map<ApiUser>(request);
+                mappedPelicanUser.PasswordHash = mappedUser.PasswordHash;
+                try
+                {
+                    await _pelicanRepository.AddAsync(mappedPelicanUser);
+
+                }
+                catch (Exception ex)
+                {
+
+                    throw;
+                }
+            }
+
+            await _logService.InsertUserActivityLog(new UserActivityLogDto
+            {
+                UserId = operatorId,
+                NewValues = mappedUser.UserName,
+                UserActivityLogTypeId = (type == UserType.Pelican) ? ActivityLogType.CreatePelicanUser : ActivityLogType.CreateTeriageUser
+            });
 
             return new ResponseDto<bool> { IsSuccessFull = true, Message = ErrorsMessages.OperationSuccessful };
         }
 
         public async Task<ResponseDto<bool>> DeleteUser(int userID, Guid operatorId)
         {
-            var user = await _identityServer.Get(userID);
+            var user = await _identityServerRepository.Get(userID);
             if (user == null)
             {
                 return new ResponseDto<bool> { IsSuccessFull = false, Message = ErrorsMessages.NotFound };
             }
-            await _identityServer.Remove(user);
-
-            var units = await _pelicanRepository.GetUserUnits(user.UserName);
-            if(units != null)
-            { 
-                foreach (var item in units)
-                {
-                    await _unitGenericRepository.Remove(item);
-                }
-            }
-
-            var permisssions = await _pelicanRepository.GetUserPermissions(user.UserName);
-            if (permisssions != null)
+            await _identityServerRepository.Remove(user);
+            var userPelcian = await _pelicanRepository.GetByUsername(user.UserName);
+            if (userPelcian != null)
             {
-                foreach (var item in permisssions)
-                {
-                    await _permissionGenericRepository.Remove(item);
-                }
-            }
-            await _logService.InsertUserActivityLog(new UserActivityLogDto { UserId = operatorId, NewValues = user.UserName, UserActivityLogTypeId = ActivityLogType.DeletePelicanUser });
 
+                await _pelicanRepository.Remove(userPelcian);
+
+                var units = await _pelicanRepository.GetUserUnits(user.UserName);
+                if (units != null)
+                {
+                    foreach (var item in units)
+                    {
+                        await _unitGenericRepository.Remove(item);
+                    }
+                }
+
+                var permisssions = await _pelicanRepository.GetUserPermissions(user.UserName);
+                if (permisssions != null)
+                {
+                    foreach (var item in permisssions)
+                    {
+                        await _permissionGenericRepository.Remove(item);
+                    }
+                }
+                await _logService.InsertUserActivityLog(new UserActivityLogDto { UserId = operatorId, NewValues = user.UserName, UserActivityLogTypeId = ActivityLogType.DeletePelicanUser });
+            }
+            else
+            {
+                await _logService.InsertUserActivityLog(new UserActivityLogDto { UserId = operatorId, NewValues = user.UserName, UserActivityLogTypeId = ActivityLogType.DeleteTeriageUser });
+            }
             return new ResponseDto<bool> { IsSuccessFull = true, Message = ErrorsMessages.OperationSuccessful };
         }
 
-        public async Task<ResponseDto<IEnumerable<Domain.Entities.IdentityServer.User>>> GetPaginatedUsersList(PaginationDto request)
+        public async Task<ResponseDto<IEnumerable<Domain.Entities.IdentityServer.User>>> GetPaginatedUsersList(PaginationDto request, UserType type)
         {
-            var users = await _identityServer.GetPaginatedUsersList(request);
+            var users = await _identityServerRepository.GetPaginatedUsersList(request, type);
             return new ResponseDto<IEnumerable<Domain.Entities.IdentityServer.User>>
             {
                 IsSuccessFull = true,
@@ -142,7 +176,7 @@ namespace PelicanManagement.Application.Services.Implementations
 
         public async Task<ResponseDto<bool>> UpdateUser(int userID, UpdateIdentityUserDto request, Guid operatorId)
         {
-            var user = await _identityServer.Get(userID);
+            var user = await _identityServerRepository.Get(userID);
             if (user == null)
             {
                 return new ResponseDto<bool> { IsSuccessFull = false, Message = ErrorsMessages.NotFound };
@@ -153,31 +187,28 @@ namespace PelicanManagement.Application.Services.Implementations
                 user.PasswordHash = _passwordHasher.HashPassword(user, request.Password);
             }
             var mappedUser = _mapper.Map(request, user);
-            await _identityServer.UpdateAsync(mappedUser);
+            await _identityServerRepository.UpdateAsync(mappedUser);
 
-          
-                var permissions = await _pelicanRepository.GetUserPermissions(user.UserName);
-                if(permissions != null)
-                {
+
+            var permissions = await _pelicanRepository.GetUserPermissions(user.UserName);
+            if (permissions != null)
+            {
                 await _permissionGenericRepository.RemoveRangeAsync(permissions);
 
-                }
+            }
             if (request.PermissionIds != null)
             {
-
                 foreach (var item in request.PermissionIds)
                 {
                     var userPermission = new UserPermission { UserId = user.UserName, PermissionId = item };
                     await _permissionGenericRepository.AddAsync(userPermission);
                 }
             }
-
-           
-                var units = await _pelicanRepository.GetUserUnits(user.UserName);
-                if (units != null)
-                {
-                    await _unitGenericRepository.RemoveRangeAsync(units);
-                }
+            var units = await _pelicanRepository.GetUserUnits(user.UserName);
+            if (units != null)
+            {
+                await _unitGenericRepository.RemoveRangeAsync(units);
+            }
             if (request.UnitIds != null)
             {
                 foreach (var item in request.UnitIds)
@@ -186,8 +217,11 @@ namespace PelicanManagement.Application.Services.Implementations
                     await _unitGenericRepository.AddAsync(userUnits);
                 }
             }
-            await _logService.InsertUserActivityLog(new UserActivityLogDto { UserId = operatorId, NewValues = mappedUser.UserName, UserActivityLogTypeId = ActivityLogType.UpdatePelicanUser });
-
+            if (await _pelicanRepository.IsExist(x => x.UserName.Equals(user.UserName)))
+            {
+                await _logService.InsertUserActivityLog(new UserActivityLogDto { UserId = operatorId, NewValues = mappedUser.UserName, UserActivityLogTypeId = ActivityLogType.UpdatePelicanUser });
+            }
+            await _logService.InsertUserActivityLog(new UserActivityLogDto { UserId = operatorId, NewValues = mappedUser.UserName, UserActivityLogTypeId = ActivityLogType.UpdateTeriageUser });
             return new ResponseDto<bool> { IsSuccessFull = true, Message = ErrorsMessages.OperationSuccessful };
         }
 
@@ -195,7 +229,7 @@ namespace PelicanManagement.Application.Services.Implementations
         public async Task<ResponseDto<IdentityUserDetailDto>> GetUserDetailByUserId(string username)
         {
             IdentityUserDetailDto detailDto = new IdentityUserDetailDto();
-            var user = await _identityServer.GetByUsername(username);
+            var user = await _identityServerRepository.GetByUsername(username);
             if (user == null)
             {
                 return new ResponseDto<IdentityUserDetailDto> { IsSuccessFull = false, Message = ErrorsMessages.NotFound };
@@ -210,6 +244,21 @@ namespace PelicanManagement.Application.Services.Implementations
         {
             var user = await _pelicanRepository.GetPermissionsAndUnits();
             return new ResponseDto<PermissionsAndUnitsDto> { IsSuccessFull = true, Data = user, Message = ErrorsMessages.OperationSuccessful, Status = "SuccessFull" };
+        }
+
+        public async Task<ResponseDto<Domain.Entities.IdentityServer.User>> GetTeriageUserDetailByUserId(string username)
+        {
+            var user = await _identityServerRepository.GetByUsername(username);
+            if (user == null)
+            {
+                return new ResponseDto<Domain.Entities.IdentityServer.User> { IsSuccessFull = false, Message = ErrorsMessages.NotFound };
+            }
+            var IsTeriage = await _pelicanRepository.IsExist(x => x.UserName.Equals(username));
+            if (IsTeriage)
+            {
+                return new ResponseDto<Domain.Entities.IdentityServer.User> { IsSuccessFull = false, Message = ErrorsMessages.IsNotTeriageUser };
+            }
+            return new ResponseDto<Domain.Entities.IdentityServer.User> { IsSuccessFull = true, Data = user, Message = ErrorsMessages.OperationSuccessful, Status = "SuccessFull" };
         }
     }
 }
